requirements.txt
flask
flask-sqlalchemy
Werkzeug
flask-migrate
flask-restful
importlib-metadata
importlib-resources
ipdb
faker
sqlalchemy-serializer
flask-mail
